+++ programming-1/problem-3/problem-3/input_generator.h
@@ -4,7 +4,7 @@
 
 #include "matrix.h"
 
-matrix *generate_matrix(short max_size, int upper_limit);
-matrix *generate_matrix_array(short max_size, int upper_limit, size_t num);
+matrix *generate_matrix(size_t max_size, int upper_limit);
+matrix *generate_matrix_array(size_t max_size, int upper_limit, size_t num);
 
 int *generate_numbers_array(int upper_limit, size_t num);
